+++ examples/quickstart/tests/test_quickstart.py
@@ -48,3 +48,35 @@ def test_full_cycle(project, remote_orchestrator):
 
     # break it down
     service_instance.delete()
+
+
+def test_transient_state(project, remote_orchestrator):
+    # get connection to remote_orchestrator
+    client = remote_orchestrator.client
+
+    # setup project
+    project.compile(
+        """
+        import quickstart
+        """
+    )
+
+    # sync project and export service entities
+    remote_orchestrator.export_service_entities()
+
+    # verify the service is in the catalog
+    result = client.lsm_service_catalog_get_entity(remote_orchestrator.environment, SERVICE_NAME)
+    assert result.code == 200
+
+    service_instance = remote_orchestrator.get_managed_instance(SERVICE_NAME)
+
+    # create an instance and wait for it to be up
+    service_instance.create(
+        attributes={"router_ip": "10.1.9.17", "interface_name": "eth1", "address": "10.0.0.254/24", "vlan_id": 14},
+        wait_for_states=["creating", "up"],
+    )
+
+    service_instance.wait_for_state(state="up")
+
+    # break it down
+    service_instance.delete()